+++ NEWS.md
@@ -1,5 +1,11 @@
 # cardx 0.2.5.9002
 
+* The following functions have been renamed.
+    - `ard_continuous()` to `ard_summary()`
+    - `ard_categorical()` to `ard_tabulate()`
+    - `ard_dichotomous()` to `ard_tabulate_value()`
+    - `ard_categorical_max()` to `ard_tabulate_max()`
+
 * Added function `ard_tabulate_abnormal()` to calculate ARDs for abnormality analyses. (#310)
 
 * Adding `strata` argument to `ard_tabulate_max()`. (#445, @jtalboys)