+++ NEWS.md
@@ -1,4 +1,4 @@
-# cardx 0.2.5.9002
+# cardx 0.2.5.9003
 
 * The following functions have been renamed.
     - `ard_continuous()` to `ard_summary()`